requirements.txt
requests==2.26.0
beautifulsoup4==4.10.0
pandas==1.3.3
scikit-learn==0.24.2
Pillow==8.3.2
